pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipeline-execution"
version = "0.1.0"
description = "Production-ready FastAPI + Celery + Redis starter kit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "celery[redis]>=5.4.0,<6.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "python-json-logger>=3.3.0,<4.0.0",
  "redis>=5.2.0,<6.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "black>=25.1.0,<26.0.0",
  "ruff>=0.15.9,<0.16.0",
  "mypy>=1.15.0,<2.0.0",
  "pre-commit>=4.2.0,<5.0.0",
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.25.0,<1.0.0",
  "pytest-cov>=6.1.0,<7.0.0",
  "types-redis>=4.6.0.20241004",
]

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "ASYNC"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["app"]

[tool.setuptools.packages.find]
include = ["app*"]
